A3

+++ src/app/components/navbar/navbar.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 import { AutenticacionService } from 'src/app/services/autenticacion.service';
+import Swal from 'sweetalert2';
 
 @Component({
 	selector: 'app-navbar',
@@ -9,16 +10,28 @@ import { AutenticacionService } from 'src/app/services/autenticacion.service';
 })
 
 export class NavbarComponent implements OnInit {
+
 	loginActive: Boolean = true;
 	registerActive: Boolean = false;
 	portfolioActive: Boolean = false;
 	pageNotFoundActive: Boolean = false;
+	rutaActiva: any;
 
-	constructor(private router: Router, private autenticacionService: AutenticacionService) {}
-	
+	constructor(private router: Router, private autenticacionService: AutenticacionService) { }
+
 	ngOnInit(): void {
-		console.log("Ruta activa: ",this.router.url);
-		switch(this.router.url) {
+		this.rutaActiva = this.router.url;
+		console.log(this.rutaActiva);
+
+		console.log("Ruta activa: ", this.router.url);
+
+		switch (this.router.url) {
+			case '/portfolio': {
+				this.portfolioActive = true;
+				this.registerActive = false;
+				this.loginActive = false;
+				break;
+			}
 			case '/login': {
 				this.loginActive = true;
 				this.registerActive = false;
@@ -31,13 +44,6 @@ export class NavbarComponent implements OnInit {
 				this.loginActive = false;
 				break;
 			}
-			case '/portfolio': {
-				this.portfolioActive = true;
-				this.registerActive = false;
-				this.loginActive = false;
-				this.router.navigate(["/portfolio"]);
-				break;
-			}
 			default: {
 				this.pageNotFoundActive = true;
 				this.portfolioActive = false;
@@ -47,11 +53,26 @@ export class NavbarComponent implements OnInit {
 			}
 		}
 	}
+
 	logout( event: Event ) {
 		event.preventDefault;
-		sessionStorage.removeItem('token');
-		this.autenticacionService.removeToken();
-		console.log("Token removido con exito!", sessionStorage.getItem('token'));
-		this.router.navigate(['/login']);
-	 }
+		Swal.fire({
+			title: '¿Cerrar sesión?',
+			text: "",
+			icon: 'warning',
+			showCancelButton: true,
+			cancelButtonColor: '#747174',
+			cancelButtonText: 'Cancelar',
+			confirmButtonColor: '#00b5ff',
+			confirmButtonText: 'Confirmar'
+		}).then((result) => {
+			if (result.isConfirmed) {
+				sessionStorage.removeItem('token');
+				this.autenticacionService.removeToken();
+				console.log("Token removido, notifico desde navbar ", sessionStorage.getItem('token'));
+				this.router.navigate(['/login']);
+			}
+		}
+		)
+	}
 }
